Load css and image files from the correct dir
Detect at runtime if we are running from an installed setup or from the
source dir and load the files from the appropriate directory.
The directories are specified with the GNOME_CLOCKS_DATA_PATH and
GNOME_CLOCKS_IMAGE_PATH environment variables.
In the future it may make sense to add also HELP_DIR and LOCALE_DIR.

To be honest I am not thrilled by this approach, but it seems like the
de facto standard (d-feet, jokosher, etc do the same)

The new utils.py file contains a Dirs class that makes accessing this
values easier.

https://bugzilla.gnome.org/show_bug.cgi?id=681883