Import a copy of pycanberra and use it
Import a copy if pycanberra (https://github.com/psykoyiko/pycanberra)
in the repository and use it to play the alert sound.
I do not particularly like this solution, but canberra is not
introspectable and the alternatives are even worse (e.g. spawining
canberra-gtk-play in the background or writing our own C module).