Pause expensive UI updates when the UI is hidden
The timer and the stopwatch use a lot of CPU time because
they update their UI frequently. To improve this, catch the map/unmap
signals of their widgets and stop updating them while they are not
visible (mapped).

Also add methods for adding/removing the timeout to Stopwatch, to
unify the code of Stopwatch and Timer.