DigitalClock: Set last_sunrise/sunset on update
This was done in the wrong place, resulting in a complete update of
every world clock, every second. This caused a pretty high
CPU usage when using a few world clocks.

Also use a new _standalone_is_new attribute for DigitalClock instances
to ensure a full DigitalClock update after a StandaloneClock has been
added. (This prevents an empty StandaloneClock until the next full
update. This wasn't necessary before, as the bug already caused the
DigitalClock to be completely updated on every update call.)

https://bugzilla.gnome.org/show_bug.cgi?id=688069