Use custom cell rederer for alarms and clocks
Use a custom cell renderer instead of using a fake widget to create a
pixbuf and then using the standard pixbuf cell renderer.
This also cleans up a few visible things: the two day and night images are
now of the correct size and we avoid loading them over and over, the
rounded rectangle over the night image is lighter and the
selection cellrender has a bit of an offset also vertically.