Port build system to autotools
I really tried to stick to distutils(-extras) this time, but it is time
to give up. We need something that deals well with --prefix, propery
check for dependecies and that will easily allow us to have some C
modules if needed.

While touching the build system I also reorganized files a bit