Alarm: Change some internals
Rename AlarmItem.check_expired() to AlarmItem.tick() and change its
function a bit. tick() updates the state and expiration time of the
AlarmItem and rings or stops the alarm if needed.

This commit introduces the ability to stop the alarm after a defined
duration. Notice that this currently doesn't make much sense because
the alarm sound only rings a few seconds for now.