Vala rewrite
Python has served us well for prototyping the app, but we are now
switching to Vala to overcome some problems (in particular lack of
proper libcanberra bindings) and to take advantage of some other things
(libgd, egglistbox, gtk resources) not readily available in python.
This rewrite is however much more than a straight port: many of the
internals have improved, we dropped the clutter dependency, we make
larger use of .ui files and much more. I also decided to start
following the GNOME version numbering