Keep the alarm ringing
Use canberra play_full to keep the alarm ringing. Default duration is
now 3 minutes. I am not sure if it is a good value and if we should
make this configurable.
Also, there is a small pause each time the sound loop, I do not know if
this depends on the sound or if it is due to requeing the idle.