prevent creating clocks for locations with no timezone.
Here, by creating a new clock for a defined and matched location the OK button
of the dialog remains disabled, if the time zone for this location is not
known, so the user cannot add such a broken clock. In this case a
warning message is produced.
Fixes https://bugzilla.gnome.org/show_bug.cgi?id=699311