Fix compilation warnins by adding AC_PROG_CC_STDC macro
The LTR mark is added to correct the time strings in RTL languges.
Moreover the AC_PROG_CC_STDC macro was added to configure.ac to
check for the existence of the ISO Standard C support. This allows the
usage of unicode character without the compiler warnings. Like in the case
of LTR mark "\u200E".