Disabling multiple alarm for the same time
Connect signals for state(value) change of spinbuttons and active_switch in
alarmsetupdialog.ui file and for day-toggle buttons in SetupDialog class.
Add GtkRevealer in ui file and add one GtkLabel as its child and initialize
it not to reveal child. If any signal emits then store dialogbox information
in temporary alarm item and check it with all the alarms(passed as an argument
to SetupDialog class constructor) and check if there exists any alarm for
the same time and it is activated. If there is a duplicate alarm then reveal
the warning label and make done button insensitive otherwise make it sensitive.

https://bugzilla.gnome.org/show_bug.cgi?id=697659