Reimplement the search provider to query all world locations
As mentioned in the original commit, I wondered if we wanted to just
match the saved locations or if instead we wanted to be able to query
the whole libgweather database. After a bit of discussion it turns
out we want the latter.
This patch also makes the initial search async since it could be slower
now.