Fix incorrect day label on Jan 1st or Jan 2nd
day_label() compares the day of the year between timezones to determine
whether it should display a "Yesterday" or "Tomorrow" label on a user's
configured world clocks. It accounts for the case where it is Dec 31st
at the user's location (variable t below) but Jan 1st at a remote
location (variable d) by performing the following check:

if (d < t) {
  // If it is Dec 31st here and Jan 1st there (d = 1), then "tomorrow"
  return d == 1 ? _("Tomorrow") : _("Yesterday");

But (d < t) && (d == 1) will also be true when d = 1 and t = 2. This
causes the "Tomorrow" label to be displayed on world clocks where it
is in fact still yesterday, if the user's current date is Jan 2nd.

The same thing happens when the user's current day is Jan 1st - world
clocks that should display "Tomorrow" display "Yesterday" instead.

To fix this, we can simply account for the Jan 2nd edge case when
d < t as follows:

  return (d == 1 && t != 2) ? _("Tomorrow") : _("Yesterday");

and similar for the (d > t) case.

https://bugzilla.gnome.org/show_bug.cgi?id=722465