Port to GtkFlowBox
Reimplement the ContentView widget using FlowBox instead of IconView.
Alarm and World now use a template widget for their tiles. A lot of
code and css is removed and simplified.

This port tries to look and behave exactly the same as the old code:
dropping IconView and a lot of custom drawing code will allow easier
design changes, but such changes are not in the scope of this patch.
However FlowBox behaves slightly differently from IconView when
resizing (the space among tiles changes and this is especially
noticeable when the action bar appears)