Fix the incorrect order of timer's time labels using RTL locale
Previously, with RTL locale and when the timer was running, time
labels were displayed in incorrect order, [ss]:[mm]:[hh]. Spinbuttons
had already been forced to LTR order, but labels were not.

This commit fixes the issue by forcing the order of time labels to LTR
also when the timer is running.

Closes #743821

Signed-off-by: Tuomas Räsänen <tuomasjjrasanen@tjjr.fi>