timer: Fix incorrect minutes reset
When resetting the timer, the values for the spinbuttons are determined
from the timer starting amount in seconds. In order to get the correct
number of minutes from this value, we must first account for the number
of hours present (mod 3600) before converting from seconds to minutes
(truncated division by 60).

https://bugzilla.gnome.org/show_bug.cgi?id=774479