Activate alarms using systemd timers
This is a reworking of the way that alarms are activated, which allows
for alarms to be sounded when Clocks is closed. The way it works is like
this:

When the session is started up, a small binary is invoked which starts
gnome-clocks just enough to load the list of alarms, create transient
systemd units for them, and then exit. These units are created, updated
and destroyed when the alarm is edited in the UI.

To fire an alarm, we start the same small binary which invokes a GAction
on gnome-clocks to ring the alarm in question.

After this commit, Clocks depends on systemd in the session for its
alarm functionality.