app: Expose world clock locations to gnome-shell
Currently gnome-shell integrates with Clocks by picking up our settings
and showing the time of the "World Clocks" locations in the calendar
drop-down. However that only works when gnome-clocks is installed on
the host, but not from flatpak, as the setting is "hidden away" in a
container in that case.

To address this, export the relevant information over D-Bus instead,
so that gnome-shell can pick it up without poking at our GSettings
schema.

https://gitlab.gnome.org/GNOME/gnome-shell/issues/1158