world: Ensure sunset happens after sunrise
When a location is placed near the international day line,
calculate_sunrise_sunset() might return sunset time earlier than sunrise.
Really it means that sunset happens early next day, but it throws
everything off.

To prevent this, add one day to the sunset datetime if it's in the past
compared to the sunrise.