alarm: make new alarm active by default
Currently, alarm is not active when added. The typical behavoir or other
clock applications is to activate the alarm upon adding it.

This commit modifies the alarm object before it's added in
activate_new() making it active.

It would be nice to make the default behaviour of the bool be
true, but since the get/set methods are overridden, vala disallows the
default property.