window: Rewrite the code for keeping track of window's state
In GTK3 we use the "window-state-event" and "configure-event" signals
for keeping track of the window state. These signals are no longer
available in GTK4.

Instead, the recommended method for keeping track of the window's state
is to bind the window object's properties to gsettings.

See https://wiki.gnome.org/HowDoI/SaveWindowState