world-face: fix sorting of locations
Fixes the sorting of world clocks. Previously, the constant
Gdk.CURRENT_TIME was used, however this constant is not valid for GLib
functions, instead its literal value of 0 was used as the unix time.
This resulted in sorting being wrong in certain situations involving
daylight savings time. The new sorting mirrors the sorting done in
gnome-shell's date menu, using the actual local time to compute and
compare offsets for each timezone.

Fixes #278.