Replace 'alarm-clock-elapsed' with 'beep-beep'
This replaces 'alarm-clock-elapsed.oga' with 'alarms/beep-beep.oga',
which is the same sound but normalized to better fit with other sounds
and shortened to take less memory. It also renames and moves it to help
having more alarm sounds. The file is now a .flac file that is converted
into an .oga file at compile time.